Update README.md
The Simple Banking System project is a simple project developed using Python that manages all the basic functions of creating a new account, view account holders record, withdraws and deposit amount, balance inquiry, and more. It is easy to use and provides the simplest management of bank account and transaction.